refactor: split route-provider.ts into route, route-provider, routes-provider
