+++ src/event-router/types/index.ts
@@ -1,4 +1,5 @@
 export * from './execution-strategy.js';
 export * from './route-decider.js';
 export * from './route-provider.js';
-
+export * from './route.js';
+export * from './routes-provider.js';
